NFLTop100/requirements.txt
Django>=3.2.14
djangorestframework>=3.13.0
django-filter>=21.1
django-cors-headers>=3.10.0
